refactor: 팔로우 관련 코드 품질 개선
- FollowUserResponseDto: UserResponseDto 필드 재사용 개선 (헬퍼 메서드 추가)
- UserFollowRepositoryImpl: 코드 중복 제거 및 findUsersByFollowDirection 직접 호출
- AdminServiceImpl: status 필수 검증 추가 및 구체적인 에러 코드 사용
- ErrorCode: FOLLOW_STATUS_REQUIRED 에러 코드 추가
- AdminFollowRepositoryImpl, UserFollowRepositoryImpl: 페이징 offset 오버플로우 방지 (Math.toIntExact 사용)

+++ sharedPrompts/src/main/java/org/example/sharedprompts/domain/follow/repository/admin/AdminFollowRepositoryImpl.java
@@ -112,7 +112,7 @@ public class AdminFollowRepositoryImpl implements AdminFollowRepository {
         long total = countQ.getSingleResult();
         
         List<User> content = query
-                .setFirstResult((int) pageable.getOffset())
+                .setFirstResult(Math.toIntExact(pageable.getOffset()))
                 .setMaxResults(pageable.getPageSize())
                 .getResultList();
